Add files via upload
The updated version of the MLPM-FKNN classifier. In this method, first training data is grouped into each class and then the set of k-nearest neighbors is found from each class. Multi-local power mean vectors are then calculated for the set of nearest neighbors from each class. Next steps are the same as in the original MLPM-FKNN method.